refactor(protocol): split server.ts schema monolith into domain files
Split the 3365-line packages/protocol/src/schemas/server.ts into seven
domain files under schemas/server/ (connection, stream, activity,
control-room, session, billing, messages), re-exported by server.ts as a
thin barrel so every existing `from './schemas/server.ts'` import path is
unchanged (#6201 Tier-3).

- Pure move: contiguous line-range partition, every non-blank body line
  assigned to exactly one domain (no loss/dup/overlap).
- Cross-domain references resolve as backward-pointing value imports (the
  file was already in dependency order — 0 cycles, 0 forward const refs).
  The single forward type reference (buildPermissionRequestMessage ->
  ServerPermissionRequestMessage) is an erased `import type`, so there is
  no runtime cycle.
- NodeNext-safe: barrel keeps server.ts a file (explicit .ts-extension
  consumer imports unaffected); dist rebuilt to match.

Verified: protocol build + 340 tests, store-core/dashboard/app typecheck,
server runtime export smoke (full surface resolves through the barrel).
